# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  auth.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginOutput:
    properties:
      token:
        type: string
    type: object
host: localhost:4002
info:
  contact: {}
  description: API for a My BarberShop application
  title: My Barbershop API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login as an admin
      parameters:
      - description: login input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginOutput'
      summary: Login
      tags:
      - auth
swagger: "2.0"